pyproject.toml
[project]
name = "ai-micro-project-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.34.0",
    "json-repair>=0.50.0",
    "langfuse==2.59.7",
    "litellm[caching]>=1.76.2",
    "no-implicit-optional>=1.4",
    "omegaconf>=2.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "tenacity>=9.1.2",
    "tiktoken>=0.11.0",
    "typer>=0.17.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
aipg = "aipg:main"

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.12",
    "types-pyyaml>=6.0.12.20250822",
]

[tool.hatch.build.targets.wheel]
packages = ["aipg"]

[tool.hatch.build.targets.sdist]
include = ["aipg", "README.md", "pyproject.toml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
]
